Fix test suite package
