+++ logging/src/jvmTest/kotlin/io/spine/logging/backend/LightweightProcessorSpec.kt
@@ -24,42 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.jvm.backend.given
+package io.spine.logging.backend
 
-import com.google.errorprone.annotations.CanIgnoreReturnValue
-import io.spine.logging.jvm.MetadataKey
-import io.spine.logging.backend.Metadata
+import org.junit.jupiter.api.DisplayName
 
-/**
- * A mutable [Metadata] implementation for testing logging backends
- * and other log handling code.
- *
- * @see <a href="http://rb.gy/h75mb">Original Java code of Google Flogger</a> for historical context.
- */
-class FakeMetadata : Metadata() {
-
-    private class KeyValuePair<T : Any>(val key: MetadataKey<T>, val value: T)
-
-    private val entries = mutableListOf<KeyValuePair<*>>()
-
-    /**
-     * Adds a key/value pair to this [Metadata].
-     */
-    @CanIgnoreReturnValue
-    fun <T : Any> add(key: MetadataKey<T>, value: T): FakeMetadata {
-        entries.add(KeyValuePair(key, value))
-        return this
-    }
-
-    override fun size(): Int = entries.size
-
-    override fun getKey(n: Int): MetadataKey<Any> = entries[n].key as MetadataKey<Any>
-
-    override fun getValue(n: Int): Any = entries[n].value
-
-    override fun <T : Any> findValue(key: MetadataKey<T>): T? {
-        val entry = entries.firstOrNull { it.key == key }
-        val casted = key.cast(entry?.value) // It is safe to pass `null` here.
-        return casted
-    }
-}
+@DisplayName("`LightweightProcessor` should")
+internal class LightweightProcessorSpec :
+    MetadataProcessorSpec(MetadataProcessor::getLightweightProcessor)
