+++ logging/src/jvmTest/kotlin/io/spine/logging/backend/MetadataHandlerSpec.kt
@@ -24,23 +24,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.jvm.backend
+package io.spine.logging.backend
 
 import com.google.common.base.Joiner
 import com.google.common.collect.Iterators
 import io.kotest.matchers.shouldBe
-import io.spine.logging.backend.Metadata
-import io.spine.logging.backend.MetadataHandler
-import io.spine.logging.backend.MetadataProcessor
+import io.spine.logging.backend.given.FakeMetadata
 import io.spine.logging.jvm.MetadataKey
-import io.spine.logging.jvm.backend.given.FakeMetadata
 import io.spine.logging.jvm.repeatedKey
 import io.spine.logging.jvm.singleKey
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 
 /**
- * Tests for [io.spine.logging.backend.MetadataHandler].
+ * Tests for [MetadataHandler].
  *
  * @see <a href="https://github.com/google/flogger/blob/cb9e836a897d36a78309ee8badf5cad4e6a2d3d8/api/src/test/java/com/google/common/flogger/backend/MetadataHandlerTest.java">
  *     Original Java code of Google Flogger</a> for historical context.
