+++ logging/src/jvmTest/kotlin/io/spine/logging/backend/MetadataKeyValueHandlersSpec.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.jvm.backend
+package io.spine.logging.backend
 
 import io.kotest.matchers.collections.shouldContainExactly
 import io.kotest.matchers.collections.shouldContainExactlyInAnyOrder
@@ -39,7 +39,7 @@ import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 
 /**
- * Tests for [io.spine.logging.backend.MetadataKeyValueHandlers].
+ * Tests for [MetadataKeyValueHandlers].
  *
  * @see <a href="https://github.com/google/flogger/blob/cb9e836a897d36a78309ee8badf5cad4e6a2d3d8/api/src/test/java/com/google/common/flogger/backend/MetadataKeyValueHandlersTest.java">
  *     Original Java code of Google Flogger</a> for historical context.
