+++ logging/src/jvmTest/kotlin/io/spine/logging/jvm/CountingRateLimiterSpec.kt
@@ -32,7 +32,7 @@ import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.types.shouldBeSameInstanceAs
 import io.spine.logging.jvm.LogContext.Key
 import io.spine.logging.jvm.RateLimitStatus.Companion.DISALLOW
-import io.spine.logging.jvm.backend.given.FakeMetadata
+import io.spine.logging.backend.given.FakeMetadata
 import io.spine.logging.jvm.given.FakeLogSite
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
