+++ logging/src/jvmTest/kotlin/io/spine/logging/jvm/DurationRateLimiterSpec.kt
@@ -32,11 +32,10 @@ import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.types.shouldBeSameInstanceAs
 import io.spine.logging.jvm.LogContext.Key
 import io.spine.logging.jvm.RateLimitStatus.Companion.DISALLOW
-import io.spine.logging.jvm.backend.given.FakeMetadata
+import io.spine.logging.backend.given.FakeMetadata
 import io.spine.logging.jvm.given.FakeLogSite
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeUnit.MILLISECONDS
-import kotlin.time.DurationUnit
 import kotlin.time.DurationUnit.SECONDS
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
