+++ logging/src/jvmTest/kotlin/io/spine/logging/jvm/LogContextSpec.kt
@@ -41,12 +41,12 @@ import io.spine.logging.backend.probe.MemoizingLoggerBackend
 import io.spine.logging.jvm.DurationRateLimiter.Companion.newRateLimitPeriod
 import io.spine.logging.jvm.LogContext.Companion.specializeLogSiteKeyFromMetadata
 import io.spine.logging.jvm.LogContext.Key
-import io.spine.logging.jvm.backend.given.FakeMetadata
-import io.spine.logging.jvm.backend.given.shouldContain
-import io.spine.logging.jvm.backend.given.shouldContainInOrder
-import io.spine.logging.jvm.backend.given.shouldHaveSize
-import io.spine.logging.jvm.backend.given.shouldNotContain
-import io.spine.logging.jvm.backend.given.shouldUniquelyContain
+import io.spine.logging.backend.given.FakeMetadata
+import io.spine.logging.backend.given.shouldContain
+import io.spine.logging.backend.given.shouldContainInOrder
+import io.spine.logging.backend.given.shouldHaveSize
+import io.spine.logging.backend.given.shouldNotContain
+import io.spine.logging.backend.given.shouldUniquelyContain
 import io.spine.logging.jvm.context.Tags
 import io.spine.logging.jvm.given.ConfigurableLogger
 import io.spine.logging.jvm.given.FakeLogSite
