+++ logging/src/jvmTest/kotlin/io/spine/logging/jvm/SamplingRateLimiterSpec.kt
@@ -34,7 +34,7 @@ import io.kotest.matchers.shouldNotBe
 import io.spine.logging.jvm.LogContext.Key
 import io.spine.logging.jvm.RateLimitStatus.Companion.DISALLOW
 import io.spine.logging.backend.Metadata
-import io.spine.logging.jvm.backend.given.FakeMetadata
+import io.spine.logging.backend.given.FakeMetadata
 import io.spine.logging.jvm.given.FakeLogSite
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
